add: ウィンドウのスケールアニメーション

+++ Assets/Scripts/ObjectAnimator/AnimationTestTrigger.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class AnimationTestTrigger : MonoBehaviour
 {
     public AnimationObserver observer;
-    public int testValue = 2;
+    public float testValue = 2;
 
     [ContextMenu("Trigger Animation")]
     public void Test()
